+++ benchmarks/scripts/benchmark_missing_omics.jl
@@ -143,7 +143,7 @@ for missing_pct in MISSING_PERCENTAGES
             from_layer_name="geno",
             to_layer_name="omics",
             equation="omics = intercept + geno",
-            omics_name=["omic$i" for i in 1:n_omics],
+            traits=["omic$i" for i in 1:n_omics],
             method="BayesC",
             estimatePi=true
         ),
@@ -151,7 +151,7 @@ for missing_pct in MISSING_PERCENTAGES
             from_layer_name="omics",
             to_layer_name="phenotypes",
             equation="phenotypes = intercept + omics",
-            phenotype_name=["trait1"],
+            traits=["trait1"],
             method="BayesC",
             activation_function="linear"
         )