+++ benchmarks/scripts/benchmark_performance.jl
@@ -60,7 +60,7 @@ function run_benchmark(chain_length::Int, burnin::Int; n_runs::Int=3)
                 from_layer_name="geno",
                 to_layer_name="omics",
                 equation="omics = intercept + geno",
-                omics_name=["omic$j" for j in 1:10],
+                traits=["omic$j" for j in 1:10],
                 method="BayesC",
                 estimatePi=true
             ),
@@ -68,7 +68,7 @@ function run_benchmark(chain_length::Int, burnin::Int; n_runs::Int=3)
                 from_layer_name="omics",
                 to_layer_name="phenotypes",
                 equation="phenotypes = intercept + omics",
-                phenotype_name=["trait1"],
+                traits=["trait1"],
                 method="BayesC",
                 activation_function="linear"
             )