+++ test/test_nnmm_basic.jl
@@ -34,7 +34,7 @@ using LinearAlgebra
             from_layer_name="geno",
             to_layer_name="omics",
             equation="omics = intercept + geno",
-            omics_name=["o1"]
+            traits=["o1"]
         )
         @test eq.from_layer_name == "geno"
         @test eq.to_layer_name == "omics"
@@ -71,14 +71,14 @@ using LinearAlgebra
                 from_layer_name="geno",
                 to_layer_name="omics",
                 equation="omics = intercept + geno",
-                omics_name=["omic1", "omic2", "omic3"],
+                traits=["omic1", "omic2", "omic3"],
                 method="BayesC"
             ),
             Equation(
                 from_layer_name="omics",
                 to_layer_name="phenotypes",
                 equation="phenotypes = intercept + omics",
-                phenotype_name=["trait1"],
+                traits=["trait1"],
                 method="BayesC"
             )
         ]